refactor: extract LayoutFooter from Layout component
Extract the footer-related functionality from the Layout component into a
separate, reusable LayoutFooter component. This new LayoutFooter component
accepts a FooterBody prop, enhancing flexibility and reusability for
various use cases.
