+++ lib/types/bases/ui.ts
@@ -28,6 +28,7 @@ export interface TypesReactChildren {
   sidebarButton: Types['children'];
   checkBox: Types['children'];
   footerButtons: Types['children'];
+  footerSidebar: Types['children'];
   headerButtons: Types['children'];
   headerIcons: Types['children'];
   nestedModal: Types['children'];
@@ -53,6 +54,8 @@ export interface TypesUi {
   isConditionalRendering: boolean;
   enterDuration: DURATION;
   leaveDuration: DURATION;
+  transitionFrom: string;
+  transitionTo: string;
   isPortal: boolean;
   gradientType: GRADIENT_TYPE;
   gradientPosition: GRADIENT_POSITION;
